Still the same as earlier, just added a few things
The MainMapController now looks more like the MainGameController.
ASHER! I do not understand your motives behind the absract
implementation of entities/Subject, and the IObsever pattern. I would
have liked to use it, but yours does not make complete sense to me. And
one more thing, I changed the TDMap class significantly again.

One thing that is left major is the addition of toggle to switch between
the panes.

The "Reinitialize the Map" button now works but leads to major glitches,
that need to be resolved.