Big changes (Towers Et Map)
I made it so Towers snap into place on the map. I changed the Map
slightly. Rather than each point corresponding to a grid of int[][], it
corresponds to a grid of MapTiles[][] which are identical (have an int
value) except they also have a Tower object. This is so that later we
can see if a tower occupies one of the positions.